[refactor] 메뉴 가격 데이터 타입을 Long 으로 통일

+++ src/main/java/camp/woowak/lab/menu/domain/Menu.java
@@ -34,7 +34,7 @@ public class Menu {
 	private String name;
 
 	@Column(nullable = false)
-	private Integer price;
+	private Long price;
 
 	@Column(nullable = false)
 	private Long stockCount;
@@ -43,7 +43,7 @@ public class Menu {
 	private String imageUrl;
 
 	public Menu(Store store, MenuCategory menuCategory, String name,
-				Integer price, Long stockCount, String imageUrl
+				Long price, Long stockCount, String imageUrl
 	) {
 		MenuValidator.validate(store, menuCategory, name, price, stockCount, imageUrl);
 		this.store = store;
